Reworked Delete and Put endpoints in CarController, added functionality to CarResponseDTO and Mapper

+++ carService/src/main/java/lt/techin/controller/CarController.java
@@ -3,14 +3,15 @@ package lt.techin.controller;
 import jakarta.validation.Valid;
 import lt.techin.dto.*;
 import lt.techin.model.Car;
+import lt.techin.model.CarStatus;
 import lt.techin.service.CarService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api")
@@ -24,22 +25,23 @@ public class CarController {
     }
 
     @GetMapping("/cars")
-    public ResponseEntity<List<CarDTO>> getCars() {
-        return ResponseEntity.ok(CarMapper.toCarDTOList(carService.findAllCars()));
+    public ResponseEntity<List<CarRequestDTO>> getCars() {
+        return ResponseEntity.ok(CarRequestMapper.toCarDTOList(carService.findAllCars()));
     }
 
     @GetMapping("/cars/{id}")
-    public ResponseEntity<CarDTO> getCar(@PathVariable long id) {
-        Optional<Car> foundCar = carService.findCarById(id);
-        if (foundCar.isEmpty()) {
-            return ResponseEntity.notFound().build();
+    public ResponseEntity<?> getCar(@PathVariable long id) {
+        if (!carService.existsCarById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Car with this ID does not exist!");
         }
-        return ResponseEntity.ok(CarMapper.toCarDTO(foundCar.get()));
+        Car foundCar = carService.findCarById(id).get();
+
+        return ResponseEntity.ok(CarRequestMapper.toCarDTO(foundCar));
     }
 
     @PostMapping("/cars")
-    public ResponseEntity<?> addCar(@Valid @RequestBody CarDTO carDTO) {
-        Car car = CarMapper.toCar(carDTO);
+    public ResponseEntity<?> addCar(@Valid @RequestBody CarRequestDTO carRequestDTO) {
+        Car car = CarRequestMapper.toCar(carRequestDTO);
 
         Car savedCar = carService.saveCar(car);
 
@@ -48,35 +50,39 @@ public class CarController {
                                 .path("/{id}")
                                 .buildAndExpand(savedCar.getId())
                                 .toUri())
-                .body(CarMapper.toCarDTO(savedCar));
+                .body(CarRequestMapper.toCarDTO(savedCar));
     }
 
     @PutMapping("/cars/{id}")
-    public ResponseEntity<?> updateCar(@PathVariable long id, @Valid @RequestBody CarDTO carDTO) {
+    public ResponseEntity<?> updateCar(@PathVariable long id, @Valid @RequestBody CarRequestDTO carRequestDTO) {
         if (carService.existsCarById(id)) {
             Car carFromDB = carService.findCarById(id).get();
 
-            CarMapper.updateCarFromDTO(carFromDB, carDTO);
+            CarRequestMapper.updateCarFromDTO(carFromDB, carRequestDTO);
 
             carService.saveCar(carFromDB);
 
-            return ResponseEntity.ok(carDTO);
+            return ResponseEntity.ok(CarResponseMapper.toCarResponseDTO(carFromDB));
         }
 
-        Car savedCar = carService.saveCar(CarMapper.toCar(carDTO));
+        Car savedCar = carService.saveCar(CarRequestMapper.toCar(carRequestDTO));
 
         return ResponseEntity.created(
                         ServletUriComponentsBuilder.fromCurrentRequest()
                                 .replacePath("/api/cars/{id}")
                                 .buildAndExpand(savedCar.getId())
                                 .toUri())
-                .body(CarMapper.toCarDTO(savedCar));
+                .body(CarRequestMapper.toCarDTO(savedCar));
     }
 
     @DeleteMapping("/cars/{id}")
-    public ResponseEntity<Void> deleteCar(@PathVariable long id) {
+    public ResponseEntity<?> deleteCar(@PathVariable long id) {
         if (!carService.existsCarById(id)) {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Car with this ID does not exist!");
+        }
+
+        if (carService.findCarById(id).get().getStatus().equals(CarStatus.valueOf("RENTED"))) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Car with this ID is currently rented!");
         }
 
         carService.deleteCarById(id);
